Please update the form to show these sections only in the admin mode: 1. Work 2. Home 3. Other
For all these sections give me a field where I can input the address and once I click Done, it becomes permanent and is stored in the Database.

For the Other section, provide me with a text box too for me to put categories of the address such as Cafe, Museum, etc. And a Add More button too for this section to add more than one address.