You remember I talked about 2 versions/views of this website, one admin, and one from my friend's POV.
How do I access that?

I have an idea that Add a Login Page.
For admin, I want the credentials to be:
Username- admin
Password- hehe

For my friend, I'll be sharing the credentials with her:
Username: masti
Password: kamyaabi

Since it is not going to change, can we get a way to incorporate this?